pyproject.toml
[tool.poetry]
name = "siestai"
version = "0.1.0"
description = "Siestai is a tool for creating and managing your AI agents."
authors = ["orhanors <orsorhan1@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.13,<4.0"
langchain = "^0.3.26"
langgraph = "^0.5.0"
langchain-openai = "^0.3.27"
sqlalchemy = "^2.0.41"
psycopg2-binary = "^2.9.10"
alembic = "^1.16.2"
python-dotenv = "^1.1.1"
pgvector = "^0.4.1"
requests = "^2.32.4"
asyncpg = "^0.30.0"
rich = "^14.0.0"
graphiti-core = "^0.14.0"
beautifulsoup4 = "^4.13.4"
nats-py = "^2.10.0"
fastapi = "^0.116.0"
uvicorn = "^0.35.0"
faststream = {extras = ["nats"], version = "^0.5.47"}
langchain-community = "^0.3.27"
langchain-ollama = "^0.3.6"
langchain-google-genai = "^2.1.8"
tavily-python = "^0.7.10"
mcp = "^1.12.2"
langgraph-checkpoint-postgres = "^2.0.23"
langchain-anthropic = "^0.3.17"

[tool.poetry.scripts]
memory-api = "start_memory_api:main"
database-ingest = "start_database_ingest:main"
kg-ingest = "start_kg_ingest:main"
research-agent = "start_research_agent:main"
c9s-agent = "start_c9s_agent:main"
c9s-terminal = "start_c9s_terminal:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"